[parquet] Remove old parquet reader
