--- paimon-format/src/main/java/org/apache/paimon/format/parquet/reader/FixedLenBytesColumnReader.java
@@ -1,60 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.format.parquet.reader;
-
-import org.apache.paimon.data.columnar.writable.WritableColumnVector;
-
-import org.apache.parquet.column.ColumnDescriptor;
-import org.apache.parquet.column.page.PageReadStore;
-import org.apache.parquet.io.api.Binary;
-import org.apache.parquet.schema.PrimitiveType;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-
-/** Fixed length bytes {@link ColumnReader}, just for Decimal. */
-public abstract class FixedLenBytesColumnReader<VECTOR extends WritableColumnVector>
-        extends AbstractColumnReader<VECTOR> {
-
-    protected final int precision;
-
-    public FixedLenBytesColumnReader(
-            ColumnDescriptor descriptor, PageReadStore pageReadStore, int precision)
-            throws IOException {
-        super(descriptor, pageReadStore);
-        checkTypeName(PrimitiveType.PrimitiveTypeName.FIXED_LEN_BYTE_ARRAY);
-        this.precision = precision;
-    }
-
-    protected void skipDataBinary(int len) {
-        skipDataBuffer(len);
-    }
-
-    protected Binary readDataBinary(int len) {
-        ByteBuffer buffer = readDataBuffer(len);
-        if (buffer.hasArray()) {
-            return Binary.fromConstantByteArray(
-                    buffer.array(), buffer.arrayOffset() + buffer.position(), len);
-        } else {
-            byte[] bytes = new byte[len];
-            buffer.get(bytes);
-            return Binary.fromConstantByteArray(bytes);
-        }
-    }
-}
