+++ paimon-format/src/main/java/org/apache/paimon/format/parquet/reader/ParquetDictionary.java
@@ -21,7 +21,16 @@ package org.apache.paimon.format.parquet.reader;
 import org.apache.paimon.data.Timestamp;
 import org.apache.paimon.data.columnar.Dictionary;
 
-import static org.apache.paimon.format.parquet.reader.TimestampColumnReader.decodeInt96ToTimestamp;
+import org.apache.parquet.io.api.Binary;
+
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+
+import static org.apache.paimon.format.parquet.writer.ParquetRowDataWriter.JULIAN_EPOCH_OFFSET_DAYS;
+import static org.apache.paimon.format.parquet.writer.ParquetRowDataWriter.MILLIS_IN_DAY;
+import static org.apache.paimon.format.parquet.writer.ParquetRowDataWriter.NANOS_PER_MILLISECOND;
+import static org.apache.paimon.format.parquet.writer.ParquetRowDataWriter.NANOS_PER_SECOND;
+import static org.apache.paimon.utils.Preconditions.checkArgument;
 
 /** Parquet dictionary. */
 public final class ParquetDictionary implements Dictionary {
@@ -60,5 +69,30 @@ public final class ParquetDictionary implements Dictionary {
     @Override
     public Timestamp decodeToTimestamp(int id) {
         return decodeInt96ToTimestamp(true, dictionary, id);
+    }
+
+    public static Timestamp decodeInt96ToTimestamp(
+            boolean utcTimestamp, org.apache.parquet.column.Dictionary dictionary, int id) {
+        Binary binary = dictionary.decodeToBinary(id);
+        checkArgument(binary.length() == 12, "Timestamp with int96 should be 12 bytes.");
+        ByteBuffer buffer = binary.toByteBuffer().order(ByteOrder.LITTLE_ENDIAN);
+        return int96ToTimestamp(utcTimestamp, buffer.getLong(), buffer.getInt());
+    }
+
+    public static Timestamp int96ToTimestamp(boolean utcTimestamp, long nanosOfDay, int julianDay) {
+        long millisecond = julianDayToMillis(julianDay) + (nanosOfDay / NANOS_PER_MILLISECOND);
+
+        if (utcTimestamp) {
+            int nanoOfMillisecond = (int) (nanosOfDay % NANOS_PER_MILLISECOND);
+            return Timestamp.fromEpochMillis(millisecond, nanoOfMillisecond);
+        } else {
+            java.sql.Timestamp timestamp = new java.sql.Timestamp(millisecond);
+            timestamp.setNanos((int) (nanosOfDay % NANOS_PER_SECOND));
+            return Timestamp.fromSQLTimestamp(timestamp);
+        }
+    }
+
+    private static long julianDayToMillis(int julianDay) {
+        return (julianDay - JULIAN_EPOCH_OFFSET_DAYS) * MILLIS_IN_DAY;
     }
 }
