+++ paimon-format/src/main/java/org/apache/paimon/format/parquet/reader/VectorizedDeltaBinaryPackedReader.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.format.parquet.newreader;
+package org.apache.paimon.format.parquet.reader;
 
 import org.apache.paimon.data.columnar.writable.WritableByteVector;
 import org.apache.paimon.data.columnar.writable.WritableColumnVector;
@@ -34,10 +34,6 @@ import org.apache.parquet.io.ParquetDecodingException;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
-
-/* This file is based on source code from the Spark Project (http://spark.apache.org/), licensed by the Apache
- * Software Foundation (ASF) under the Apache License, Version 2.0. See the NOTICE file distributed with this work for
- * additional information regarding copyright ownership. */
 
 /**
  * An implementation of the Parquet DELTA_BINARY_PACKED decoder that supports the vectorized
