+++ paimon-format/src/main/java/org/apache/paimon/format/parquet/reader/VectorizedDeltaLengthByteArrayReader.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.format.parquet.newreader;
+package org.apache.paimon.format.parquet.reader;
 
 import org.apache.paimon.data.columnar.heap.HeapIntVector;
 import org.apache.paimon.data.columnar.writable.WritableBytesVector;
@@ -27,10 +27,6 @@ import org.apache.parquet.io.ParquetDecodingException;
 import java.io.EOFException;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-
-/* This file is based on source code from the Spark Project (http://spark.apache.org/), licensed by the Apache
- * Software Foundation (ASF) under the Apache License, Version 2.0. See the NOTICE file distributed with this work for
- * additional information regarding copyright ownership. */
 
 /**
  * An implementation of the Parquet DELTA_LENGTH_BYTE_ARRAY decoder that supports the vectorized
